backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
pyodbc==5.2.0
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.10
PyJWT==2.9.0
langdetect==1.0.9
waitress==3.0.2
psutil==6.1.0
huggingface-hub==0.26.2
